Allow exclusive full screen with custom resolutions
Previously the plugin would always use the primary monitor's resolution
for exclusive full screen mode.
With this change, the WindowWidth and WindowHeight options can be used
to configure the resolution for both the windowed and exclusive full
screen window modes.

If the exclusive full screen window size is less then the primary
monitor's resolution, the OS will change the monitor's display
resolution to whatever the application requests.
This can be useful in some scenarios, e.g. making a 4K (3840x2160)
monitor run the game in exclusive full screen at 1920x1080.